+++ .gitignore
@@ -7,3 +7,5 @@ src/.env.production
 
 # ignore package-lock - it is not used in downstream projects
 package-lock.json
+
+depgraphs
