+++ src/lib/permissions/pages/client/pagePermissionsHttpClient.ts
@@ -1,7 +1,7 @@
 import { DELETE, GET, POST } from '../../../../adapters/http';
 import type { PageMeta, PageMetaWithPermissions, PagesRequest } from '../../../pages/interfaces';
 import { AbstractPermissionsApiClient } from '../../clients/abstractApiClient.class';
-import type { PermissionCompute, PermissionResource, Resource } from '../../interfaces';
+import type { PermissionCompute, PermissionResource, Resource } from '../../core/interfaces';
 import type {
   AssignablePagePermissionGroups,
   AssignedPagePermission,
